requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
